Fix algorithm to pass all the tests
Now, aside from passing the tests, we're following and optimizing inner
functions in a smarter way (only when we need to). This has sped up *a
lot* the compilation time. Now, we're good, although we still need to
fix some issues with the core of the algorithm regarding to uncovered
common use cases.