Initialize optimized variables in order
This patch introduces a post-order traversal to respect the ordering of the optimized variables, our previous pre-order traversal is not able to guarantee the same program behaviour.

The pre-order traversal will be removed in the future with a proper parent analysis, which is required for the pruning. Our current solution is inefficient, but works.